+++ update-api/config.php
@@ -24,3 +24,4 @@ define('PLUGINS_DIR', BASE_DIR . '/storage/plugins');
 define('THEMES_DIR', BASE_DIR . '/storage/themes');
 define('BLACKLIST_DIR', BASE_DIR . '/storage');
 define('LOG_DIR', BASE_DIR . '/storage/logs');
+define('DB_FILE', BASE_DIR . '/storage/updater.sqlite');